add the ability to disable a plugin in RestartableJenkinsRule
it was not easy to write a PluginManager that loaded dependencies
correctly for Jars for use in RestartableJenkinsRule that supported
plugin disabling.

With this change UnitTestSupportingPluginManager works just like a
normal PluginManager *and* support the logic needed for maven test scope
dependencies (but without the run tests as fast as possible used by
TestPluginManager)
