+++ src/test/java/org/jvnet/hudson/test/RestartableJenkinsRuleTest.java
@@ -2,12 +2,34 @@ package org.jvnet.hudson.test;
 
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeThat;
 
+import java.io.File;
+import java.nio.file.CopyOption;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicInteger;
+
 import org.junit.Rule;
 import org.junit.Test;
+import org.jvnet.hudson.test.recipes.WithPlugin;
+import org.jvnet.hudson.test.recipes.WithPluginManager;
+import org.kohsuke.stapler.HttpResponse;
+
+import hudson.PluginManager;
+import hudson.PluginWrapper;
+import hudson.model.UpdateCenter.UpdateCenterJob;
+import hudson.util.IOUtils;
+
+import jenkins.model.Jenkins;
 
 public class RestartableJenkinsRuleTest {
 
@@ -51,5 +73,36 @@ public class RestartableJenkinsRuleTest {
                 s -> {
                     assertEquals(port.get(), portReuse.j.getURL().getPort());
                 });
+    }
+
+    @Test
+    @WithPluginManager(UnitTestSupportingPluginManager.class)
+    public void pluginsCanBeDisabled() throws Exception {
+        final String pluginId = "display-url-api";
+        noPortReuse.then(jr -> {
+            System.out.println(WarExploder.getExplodedDir());
+            Path srcLdap = new File(WarExploder.getExplodedDir(), "WEB-INF/detached-plugins/"+pluginId+".hpi").toPath();
+            Path dstLdap = new File(jr.jenkins.pluginManager.rootDir, pluginId+".jpi").toPath();
+            Files.createDirectories(dstLdap.getParent());
+            Files.copy(srcLdap, dstLdap);
+            //jr.getPluginManager().doCheckUpdatesServer();
+            //jr.getPluginManager().install(Collections.singletonList("cvs"),true);
+            System.out.println("hello");
+        });
+        
+        noPortReuse.then(jr -> {
+            Jenkins j = jr.jenkins;
+            PluginManager pm = j.getPluginManager();
+            PluginWrapper plugin = pm.getPlugin(pluginId);
+            assertNotNull("plugin should not be null", plugin);
+            plugin.disable();
+        });
+
+        noPortReuse.then(jr -> {
+            assertFalse("workflow-job is not enabled",
+                        jr.jenkins.getPluginManager().getPlugin(pluginId).isEnabled());
+            assertFalse("workflow-job should not be active",
+                        jr.jenkins.getPluginManager().getPlugin(pluginId).isActive());
+        });
     }
 }
